BlogManager/requirements.txt
Flask==2.3.2
mysql-connector-python==8.0.33
pyngrok==5.2.2
python-dotenv==1.0.0
requests==2.31.0
